Create ATM.java
ATM INTERFACE

We have all come across ATMs in our cities and it is built on Java. This complex project consists of Tive different dlasses and is a console-basell application When the system starts the user is prompted with user id and user sin. Un entering the details successfully, then ATM functional des are unlocked. The project allows to perform following operations

1. Transactions History
2. Withdraw
3.Deposit
4.Transfer
5.Quit